Cleanup on the federalization service
