+++ ActionHandler/CatalogiHandler.php
@@ -4,14 +4,15 @@ namespace OpenCatalogi\OpenCatalogiBundle\ActionHandler;
 
 use OpenCatalogi\OpenCatalogiBundle\Service\CatalogiService;
 use CommonGateway\CoreBundle\ActionHandler\ActionHandlerInterface;
+use OpenCatalogi\OpenCatalogiBundle\Service\FederalizationiService;
 
 class CatalogiHandler implements ActionHandlerInterface
 {
-    private CatalogiService $catalogiService;
+    private FederalizationiService $federalizationiService;
 
-    public function __construct(CatalogiService $catalogiService)
+    public function __construct(FederalizationiService $federalizationiService)
     {
-        $this->catalogiService = $catalogiService;
+        $this->federalizationiService = $federalizationiService;
     }
 
     /**
@@ -25,7 +26,8 @@ class CatalogiHandler implements ActionHandlerInterface
         return [
             '$id'        => 'https://example.com/person.schema.json',
             '$schema'    => 'https://json-schema.org/draft/2020-12/schema',
-            'title'      => 'CatalogiHandler'
+            'title'      => 'CatalogiHandler',
+            'description' => 'Syncs  all the know catalogi'
             ];
 
         // We don't need all of this
@@ -79,6 +81,6 @@ class CatalogiHandler implements ActionHandlerInterface
      */
     public function run(array $data, array $configuration): array
     {
-        return $this->catalogiService->catalogiHandler($data, $configuration);
+        return $this->federalizationiService->catalogiHandler($data, $configuration);
     }
 }
